+++ src/main/java/it/unibo/heavypocket/mvc/model/core/Transaction.java
@@ -1,12 +1,13 @@
 package it.unibo.heavypocket.mvc.model.core;
 
 import java.time.LocalDate;
+import java.math.BigDecimal;
 
 public interface Transaction {
  
     String getDescription();
     
-    double getAmount();
+    BigDecimal getAmount();
 
     LocalDate getDate();
 }
