+++ src/test/java/it/unibo/heavypocket/mvc/model/core/ExpenseTest.java
@@ -2,6 +2,7 @@ package it.unibo.heavypocket.mvc.model.core;
 
 import java.util.Set;
 import java.time.LocalDate;
+import java.math.BigDecimal;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Assertions;
@@ -14,7 +15,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 public class ExpenseTest {
     
     private Transaction transaction;
-    private final static double AMOUNT = 100.0;
+    private final static BigDecimal AMOUNT = BigDecimal.valueOf(100.0);
     private final static String DESCRIPTION = "TestExpense";
     private final static LocalDate DATE = LocalDate.now();
     private final static Set<Tag> TAGS = Set.of(new TagImpl("Food"));
